% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{are_dfs_equal}
\alias{are_dfs_equal}
\title{Compare two data frames for equality}
\usage{
are_dfs_equal(df1, df2)
}
\arguments{
\item{df1}{The first data frame.}

\item{df2}{The second data frame.}
}
\value{
\code{TRUE} if the data frames are identical after preparation, \code{FALSE} otherwise.
}
\description{
This internal helper function prepares two data frames for comparison by:
\enumerate{
\item Converting all columns to character type.
\item Sorting columns by name.
\item Arranging rows based on all columns.
It then checks for identity using \code{identical()}.
}
}
\keyword{internal}
