% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roster.R
\name{assign_rates}
\alias{assign_rates}
\title{Assign pay rates based on role and PhD status}
\usage{
assign_rates(data)
}
\arguments{
\item{data}{A data frame containing staff roles and information.}
}
\value{
A data frame with a new 'rate' column.
}
\description{
Assign pay rates based on role and PhD status
}
