% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow-build-database.R
\name{build_database}
\alias{build_database}
\title{Build student database}
\usage{
build_database(df, uos = NULL)
}
\arguments{
\item{df}{A data frame containing file paths and metadata}

\item{uos}{Optional. Unit of study URL or code (e.g., "BIOL1009-2024-S1C-ND-CC")}
}
\value{
A list of parsed data frames (as tibbles)
}
\description{
Build a student database from Canvas, Gradescope, and other files. This
function is limited to one file of each type. Gradescope file is optional.
}
