% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roster.R
\name{check_equal_roster}
\alias{check_equal_roster}
\title{Check if two rosters are identical}
\usage{
check_equal_roster(roster, roster2)
}
\arguments{
\item{roster}{A roster data frame.}

\item{roster2}{Another roster data frame.}
}
\value{
A logical value indicating if the rosters are identical.
}
\description{
Check if two rosters are identical
}
