% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roster.R
\name{check_equal_templates}
\alias{check_equal_templates}
\title{Check if two templates are identical}
\usage{
check_equal_templates(template, template2)
}
\arguments{
\item{template}{A template data frame.}

\item{template2}{Another template data frame.}
}
\value{
A logical value indicating if the templates are identical.
}
\description{
Check if two templates are identical
}
