% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_check_urls.R
\name{check_urls}
\alias{check_urls}
\title{Check if URLs exist - optimized for 404 detection only}
\usage{
check_urls(urls)
}
\arguments{
\item{urls}{character vector - URLs to check}
}
\value{
Character vector of non-404 URLs
}
\description{
Check if URLs exist - optimized for 404 detection only
}
