% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/academic_plans.R
\name{clean_column_names}
\alias{clean_column_names}
\title{Clean Excel Column Names}
\usage{
clean_column_names(raw_names)
}
\arguments{
\item{raw_names}{Dataframe. Raw names from Excel file.}
}
\value{
A vector of column names.
}
\description{
Internal function to process and clean column names from the Excel file.
}
