% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roster.R
\name{compare_ota_templates}
\alias{compare_ota_templates}
\title{Compare a new template with the latest logged template}
\usage{
compare_ota_templates(new_template, source_file_path, log_dir = "log/")
}
\arguments{
\item{new_template}{The new template data frame.}

\item{source_file_path}{The path to the source Excel file, used for log file naming.}

\item{log_dir}{The directory where template logs are stored.}

\item{key_cols}{A character vector of key column names for joining.}
}
\description{
Compare a new template with the latest logged template
}
