% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roster.R
\name{compare_rosters}
\alias{compare_rosters}
\title{Compare a new roster with the latest logged roster}
\usage{
compare_rosters(new_roster, log_dir = "log/")
}
\arguments{
\item{new_roster}{The new roster data frame.}

\item{log_dir}{The directory where roster logs are stored.}

\item{key_cols}{A character vector of key column names for joining.}
}
\description{
Compare a new roster with the latest logged roster
}
