% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uos.R
\name{construct_url}
\alias{construct_url}
\title{Construct URL from UoS string}
\usage{
construct_url(uos_string)
}
\arguments{
\item{uos_string}{string in format "UNIT-YEAR-SEM-TYPE-DEL"}
}
\value{
URL string
}
\description{
Construct URL from UoS string
}
\keyword{internal}
