% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_docs.R
\name{detect_canvas}
\alias{detect_canvas}
\title{Detect Canvas file format}
\usage{
detect_canvas(cols)
}
\arguments{
\item{cols}{Character vector of column names}
}
\value{
Logical indicating if columns match Canvas pattern
}
\description{
Detect Canvas file format
}
