% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect-file-types.R
\name{detect_gradescope}
\alias{detect_gradescope}
\title{Detect Gradescope file format}
\usage{
detect_gradescope(cols)
}
\arguments{
\item{cols}{Character vector of column names}
}
\value{
Logical indicating if columns match Gradescope pattern
}
\description{
Detect Gradescope file format
}
\keyword{internal}
