% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_docs.R
\name{detect_spec_cons}
\alias{detect_spec_cons}
\title{Detect Special Considerations file format}
\usage{
detect_spec_cons(cols)
}
\arguments{
\item{cols}{Character vector of column names}
}
\value{
Logical indicating if columns match Special Considerations pattern
}
\description{
Detect Special Considerations file format
}
