% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect-file-types.R
\name{detect_sres}
\alias{detect_sres}
\title{Detect SRES file format}
\usage{
detect_sres(cols)
}
\arguments{
\item{cols}{Character vector of column names}
}
\value{
Logical indicating if columns match SRES pattern
}
\description{
Detect SRES file format
}
\keyword{internal}
