% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{display_selected_columns}
\alias{display_selected_columns}
\title{Display selected columns}
\usage{
display_selected_columns(selected_cols)
}
\arguments{
\item{selected_cols}{Vector of selected column names}
}
\description{
Display selected columns
}
\keyword{internal}
