% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/special_considerations.R
\name{display_summary}
\alias{display_summary}
\title{Display summary of special considerations}
\usage{
display_summary(df)
}
\description{
Display summary of special considerations
}
\keyword{internal}
