% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.get_val_or_default}
\alias{.get_val_or_default}
\title{Get Value or Default}
\usage{
.get_val_or_default(value, default_val = "N/A")
}
\arguments{
\item{value}{The value to retrieve.}

\item{default_val}{The default value to return if 'value' is empty or NA.
Defaults to "N/A".}
}
\value{
The original value, or 'default_val' if the original is empty/NA.
}
\description{
Internal helper function to return a value if it's not empty or NA,
otherwise returns a specified default value.
Designed for scalar inputs.
}
\keyword{internal}
