% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eoi_parser_app.R
\name{EoiParserApp}
\alias{EoiParserApp}
\title{Launch EOI Parser Application}
\usage{
EoiParserApp()
}
\value{
Does not return a value; launches a Shiny application.
}
\description{
This function launches the EOI Parser Shiny application.
}
