% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extensions_gui.R
\name{extensionsGUI}
\alias{extensionsGUI}
\title{Launch Extensions GUI}
\usage{
extensionsGUI()
}
\description{
This function launches a Shiny app that allows users to browse and view
special considerations files. The app provides filtering options for
unit of study code and year.
}
