% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-extensions-viewer.R
\name{ExtensionsViewer}
\alias{ExtensionsViewer}
\title{Launch Extensions Viewer Shiny Application}
\usage{
ExtensionsViewer()
}
\value{
Does not return a value; launches a Shiny application.
}
\description{
This function launches the Shiny application for viewing extensions.
The application files are located in the 'inst/shiny/extensions_viewer' directory
of the package.
}
\examples{
\dontrun{
ExtensionsViewer()
}
}
