% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/special_considerations.R
\name{filter_spec_cons}
\alias{filter_spec_cons}
\title{Filter special considerations data}
\usage{
filter_spec_cons(df, outcomes)
}
\description{
Filter special considerations data
}
\keyword{internal}
