% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find-docs.R
\name{find_docs}
\alias{find_docs}
\title{Find relevant documents for database}
\usage{
find_docs(path = NULL)
}
\arguments{
\item{path}{Path to directory containing files to analyse}
}
\value{
A tibble
}
\description{
Scans a directory for CSV and Excel (.xlsx) files and
categorises them based on their header structure into one of several
predefined types that are often used in USYD systems:
\itemize{
\item Canvas
\item Gradescope
\item Special Considerations
\item Disability Academic Plans
}
}
\examples{
\dontrun{
# Scan current directory
files <- find_docs(".")

# Show detected file types
table(files$type)

# List unrecognised files
subset(files, type == "unknown")
}
}
