% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alt.R
\name{find_files}
\alias{find_files}
\title{Validate and categorize documentation files}
\usage{
find_files(path = NULL)
}
\arguments{
\item{path}{Path to directory containing files to analyse}
}
\value{
A tibble with the following columns:
\itemize{
\item \code{path}: Full path to the file
\item \code{modified}: File modification timestamp
\item \code{type}: Detected file type (\code{"canvas"}, \code{"gradescope"}, \code{"spec_cons"},
\code{"plans"}, or \code{"unknown"})
}
}
\description{
Scans a directory for CSV and Excel (.xlsx) files and
categorizes them based on their header structure into one of several
predefined types that are often used in USYD systems:
\itemize{
\item Canvas
\item Gradescope
\item Special Considerations
\item Disability Academic Plans
}
}
\examples{
\dontrun{
# Scan current directory
files <- find_files(".")

# Show detected file types
table(files$type)

# List unrecognized files
subset(files, type == "unknown")
}
}
