% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/special_considerations.R
\name{find_spec_cons_file}
\alias{find_spec_cons_file}
\title{Find special considerations/arrangements file}
\usage{
find_spec_cons_file(dir = NULL)
}
\arguments{
\item{dir}{Character string specifying the directory path to search. If NULL
(default), uses current working directory.}
}
\value{
Character string with full file path to most recent arrangements file
}
\description{
Find special considerations/arrangements file
}
