% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_database.R
\name{join_dataframes}
\alias{join_dataframes}
\title{Join data frames from multiple sources}
\usage{
join_dataframes(canvas_df, gradescope_df, spec_cons_df)
}
\arguments{
\item{canvas_df}{Data frame containing Canvas data}

\item{gradescope_df}{Data frame containing Gradescope data}

\item{spec_cons_df}{Data frame containing special considerations data}
}
\value{
A tibble containing combined student data with redundant columns
removed
}
\description{
Combines data frames from Canvas, Gradescope, and special
considerations into a single unified data frame. Performs left joins on SID
and removes redundant columns.
}
\keyword{internal}
