% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_gradescope.R
\name{merge_gradescope}
\alias{merge_gradescope}
\title{Merge Multiple Gradescope CSV Files}
\usage{
merge_gradescope(files)
}
\arguments{
\item{files}{Character vector of file paths to Gradescope CSV files}
}
\value{
A tibble containing merged Gradescope data
}
\description{
Merge Multiple Gradescope CSV Files
}
