% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canvas.R
\name{parse_canvas}
\alias{parse_canvas}
\title{Parse Canvas export file}
\usage{
parse_canvas(x, cols = NULL)
}
\arguments{
\item{x}{Path to Canvas export file}

\item{cols}{Optional vector of column indices to select}
}
\value{
A data frame containing parsed Canvas data
}
\description{
Parse Canvas export file
}
