% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_sc.R
\name{parse_sc}
\alias{parse_sc}
\title{Parse spec cons file}
\usage{
parse_sc(x, uos = NULL, year = NULL)
}
\arguments{
\item{x}{Path to sc file (CSV or XLSX format)}

\item{uos}{Unit of study code to filter by (optional)}

\item{year}{Year to filter by (optional)}
}
\value{
A data frame containing filtered sc data
}
\description{
Parse spec cons file
}
