% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/special_considerations.R
\name{parse_spec_cons}
\alias{parse_spec_cons}
\title{Parse special considerations file}
\usage{
parse_spec_cons(x, uos = NULL, year = NULL)
}
\arguments{
\item{x}{Path to special considerations file}

\item{uos}{Unit of study code}

\item{year}{Year of study}
}
\value{
A data frame containing parsed special considerations data
}
\description{
Parse special considerations file
}
