% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_sres.R
\name{parse_sres}
\alias{parse_sres}
\title{Parse SRES CSV files into a tibble}
\usage{
parse_sres(file)
}
\arguments{
\item{file}{Path to the SRES CSV file}
}
\value{
A tibble containing parsed SRES data
}
\description{
Parse SRES CSV files into a tibble
}
