% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uos.R
\name{parse_uos}
\alias{parse_uos}
\title{Parse unit of study webpage}
\usage{
parse_uos(url)
}
\arguments{
\item{url}{string. The URL of the unit outline page}
}
\value{
a list object with parsed unit information
}
\description{
Parse unit of study webpage
}
\keyword{internal}
