% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roster.R
\name{plot_differences}
\alias{plot_differences}
\title{Plot differences between two templates}
\usage{
plot_differences(before, after)
}
\arguments{
\item{before}{The old template data frame.}

\item{after}{The new template data frame.}

\item{key_cols}{A character vector of key column names for joining.}
}
\value{
A gt table object (invisibly).
}
\description{
Plot differences between two templates
}
