% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roster.R
\name{prepare_for_comparison}
\alias{prepare_for_comparison}
\title{Prepare a template dataframe for comparison}
\usage{
prepare_for_comparison(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
A data frame prepared for comparison.
}
\description{
Prepare a template dataframe for comparison
}
