% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prompt_user_for_columns}
\alias{prompt_user_for_columns}
\title{Prompt user to select columns}
\usage{
prompt_user_for_columns(choose)
}
\arguments{
\item{choose}{Vector of column names to choose from}
}
\value{
Vector of selected column names
}
\description{
Prompt user to select columns
}
\keyword{internal}
