% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_canvas_data}
\alias{read_canvas_data}
\title{Read Canvas data from CSV file}
\usage{
read_canvas_data(x, col_names)
}
\arguments{
\item{x}{Path to CSV file}

\item{col_names}{Vector of column names}
}
\value{
Data frame of Canvas data
}
\description{
Read Canvas data from CSV file
}
\keyword{internal}
