% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_column_names}
\alias{read_column_names}
\title{Read column names from a CSV file}
\usage{
read_column_names(x)
}
\arguments{
\item{x}{Path to CSV file}
}
\value{
Vector of column names
}
\description{
Read column names from a CSV file
}
\keyword{internal}
