% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/special_considerations.R
\name{read_spec_cons_data}
\alias{read_spec_cons_data}
\title{Read special considerations data}
\usage{
read_spec_cons_data(x, uos, year)
}
\description{
Read special considerations data
}
\keyword{internal}
