% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_database.R
\name{select_markbook_assessments}
\alias{select_markbook_assessments}
\title{Select and weight assessments for marksheet}
\usage{
select_markbook_assessments(db)
}
\arguments{
\item{db}{Data frame containing the combined student data from create_database}
}
\value{
A data frame containing only the selected assessment columns along
with student identification columns
}
\description{
Takes a database of student marks and allows users to select which assessments
should be included in the marksheet along with their respective weights.
}
