% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/special_considerations.R
\name{select_outcome_type}
\alias{select_outcome_type}
\title{Select outcome type for special considerations}
\usage{
select_outcome_type(df)
}
\description{
Select outcome type for special considerations
}
\keyword{internal}
