% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\docType{package}
\name{soles-package}
\alias{soles}
\alias{soles-package}
\title{Package dependencies}
\description{
This package requires the following packages:
\itemize{
\item dplyr: For data manipulation
\item readr: For reading CSV files
\item lubridate: For date/time handling
}
}
\author{
\strong{Maintainer}: Januar Harianto \email{januarharianto@gmail.com}

}
