% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_spawn_urls.R
\name{spawn_urls}
\alias{spawn_urls}
\title{Generate a list of URLs for a unit}
\usage{
spawn_urls(unit, year = as.character(format(Sys.Date(), "\%Y")))
}
\arguments{
\item{unit}{string - Unit of study code, e.g. 'BIOL2022'}

\item{year}{numeric - Academic year, e.g. 2024}
}
\value{
A character vector of URLs
}
\description{
Perhaps there is a better way but for now, the aim is to generate a list of
URLs for a given unit and year to check if the unit outline is available in
a later step.
}
\keyword{internal}
