% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uos.R
\name{uos}
\alias{uos}
\title{Extract Unit of Study Information}
\usage{
uos(website_url)
}
\arguments{
\item{website_url}{Character string containing the URL for a University of
Sydney unit of study page}
}
\value{
A list containing:
\itemize{
\item unit: Unit code and name
\item year: Academic year
\item semester: Semester number
\item location: Campus location
\item assessments: Data frame of assessment tasks
}
}
\description{
Retrieves unit of study information from the University of Sydney website,
including unit details and assessment schedule.
}
\examples{
\dontrun{
# Extract BIOL2022 information
biol2022 <- uos("https://www.sydney.edu.au/units/BIOL2022/2024-S2C-ND-CC")

# View assessment schedule
biol2022$assessments
}
}
