% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/special_considerations.R
\name{validate_inputs}
\alias{validate_inputs}
\title{Validate special considerations inputs}
\usage{
validate_inputs(uos, year)
}
\description{
Validate special considerations inputs
}
\keyword{internal}
