% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_links.R
\name{validate_links}
\alias{validate_links}
\title{Check if URLs contain valid unit outlines}
\usage{
validate_links(urls)
}
\arguments{
\item{urls}{Character vector of URLs to check}
}
\value{
Logical vector where TRUE indicates a valid unit outline page
}
\description{
Takes URLs from spawn_urls and checks if each contains a unit outline or
redirects to an error page.
}
