% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_outline.R
\name{view_outline}
\alias{view_outline}
\title{Extract unit of study information}
\usage{
view_outline(url)
}
\arguments{
\item{url}{string. The URL of the unit outline page.}
}
\value{
a list object
}
\description{
Extract unit of study information
}
