% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_database.R
\name{build_database}
\alias{build_database}
\title{Build student database}
\usage{
build_database(df)
}
\arguments{
\item{df}{A data frame containing file paths and metadata}
}
\value{
A list of parsed data frames (as tibbles)
}
\description{
Build a student database from Canvas, Gradescope, and other files. This
function is limited to one file of each type. Gradescope file is optional.
}
