% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roster_to_ota.R
\name{compare_rosters}
\alias{compare_rosters}
\title{Compare roster with a previous version and print differences}
\usage{
compare_rosters(df, csv = NULL)
}
\arguments{
\item{df}{A data frame, typically the output of \code{\link[=process_paycodes]{process_paycodes()}}. It must
have a \code{source_file} attribute if \code{csv} is not provided.}

\item{csv}{Optional. Path to the CSV file for comparison. If \code{NULL} (default),
the function automatically finds the latest relevant roster file in the
'logs' directory based on the \code{source_file} attribute of \code{df}.}
}
\value{
Invisibly returns the input data frame \code{df} with a \code{has_changes}
attribute set to \code{TRUE} if differences were found, and \code{FALSE} otherwise.
This allows for chaining with \code{\link[=write_paycodes_to_csv]{write_paycodes_to_csv()}}.
}
\description{
This function compares a given roster data frame with a previous version from a
CSV file. It identifies and prints added, removed, and modified rows.
}
