% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect-file-types.R
\name{detect_academic_plans}
\alias{detect_academic_plans}
\title{Detect Academic Plans file format}
\usage{
detect_academic_plans(cols)
}
\arguments{
\item{cols}{Character vector of column names}
}
\value{
Logical indicating if columns match Academic Plans pattern
}
\description{
Detect Academic Plans file format
}
\keyword{internal}
