% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{find_docs}
\alias{find_docs}
\title{Find Required Files for Exam Processing}
\usage{
find_docs(directory)
}
\arguments{
\item{directory}{Character string specifying the path to search for files}
}
\value{
A list containing three elements: \item{gradescope}{Path to the
Gradescope file} \item{canvas}{Path to the Canvas file}
\item{arrangements}{Path to the Arrangements file}
}
\description{
Searches the specified directory for required files related to exam
processing: Gradescope, Canvas, and Arrangements files.
}
\examples{
files <- find_docs("path/to/directory")
}
