% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_canvas.R
\name{parse_canvas}
\alias{parse_canvas}
\title{Parse Canvas export file}
\usage{
parse_canvas(x)
}
\arguments{
\item{x}{Path to Canvas file}
}
\value{
A list
}
\description{
Reads a Canvas export file and returns a list containg student data
and unit information. The unit information is useful for filtering
other files (e.g. academic plans) to match the Canvas unit.
}
