% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradescope.R
\name{parse_gradescope}
\alias{parse_gradescope}
\title{Parse Gradescope export file}
\usage{
parse_gradescope(x)
}
\arguments{
\item{x}{Path to Gradescope export file}
}
\value{
A data frame containing parsed Gradescope data
}
\description{
Parse Gradescope export file
}
