% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roster.R
\name{parse_roster}
\alias{parse_roster}
\title{Parse a roster from an Excel file}
\usage{
parse_roster(
  path,
  roster_sheet = 1,
  staff_sheet = "staff",
  roster_skip = 4,
  id_cols = c("week", "practical"),
  role_prefix_map = c(`^sup` = "Tutor", `^demo` = "Demonstrator")
)
}
\arguments{
\item{path}{Path to the Excel file.}

\item{roster_sheet}{Name or index of the sheet containing the roster data.}

\item{staff_sheet}{Name or index of the sheet containing the staff data.}

\item{roster_skip}{Number of rows to skip when reading the roster sheet.}

\item{id_cols}{A character vector of column names to fill downwards.}

\item{role_prefix_map}{A named character vector mapping column prefixes to roles.}
}
\value{
A processed data frame.
}
\description{
This function reads roster and staff data from an Excel file, processes it,
and returns a cleaned and summarised data frame.
}
