% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_log_level}
\alias{set_log_level}
\title{Set logging level}
\usage{
set_log_level(verbose = FALSE)
}
\arguments{
\item{verbose}{If \code{TRUE}, set log level to INFO, otherwise set to SUCCESS.}
}
\value{
Nothing
}
\description{
Set logging level
}
