% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uos.R
\name{uos}
\alias{uos}
\title{Extract unit of study information}
\usage{
uos(unit, path = NULL)
}
\arguments{
\item{unit}{string. The URL of the unit outline page or a UoS string (e.g., "DATA1001-2024-S1C-ND-CC").}

\item{path}{string. Path to save/load RDS files. If NULL, uses current working directory.}
}
\value{
a list object
}
\description{
Extract unit of study information
}
